added grpc server & handlers

+++ electronic-voting-manager/app/internal/model/voting.go
@@ -3,9 +3,9 @@ package model
 import "time"
 
 type Voting struct {
-	ID        int
+	ID        int32
 	Title     string
-	EndTime   int
+	EndTime   int64
 	Address   string
 	CreatedOn time.Time
 }
